Extract PlayerPredicate, fix QuotaCondition
