+++ LethalAchievements/Config/Conditions/AnyOfCondition.cs
@@ -5,29 +5,21 @@ using Newtonsoft.Json;
 namespace LethalAchievements.Config.Conditions;
 
 /// <summary>
-///     Evaluates to true if any of the <see cref="Conditions"/> are true.
+///     Evaluates to true if any of the <see cref="Terms"/> are true.
 /// </summary>
-[JsonConverter(typeof(TransparentConverter<AnyCondition, TaggedCondition[]>))]
-public class AnyCondition : ICondition
+public class AnyOfCondition : ICondition
 {
     /// <summary>
     ///     The conditions to evaluate.
     /// </summary>
-    public TaggedCondition[] Conditions { get; }
-    
-    /// <summary>
-    ///     Creates a new <see cref="AnyCondition"/> with the specified conditions.
-    /// </summary>
-    public AnyCondition(TaggedCondition[] conditions)
-    {
-        Conditions = conditions;
-    }
+    [JsonRequired]
+    public TaggedCondition[] Terms;
 
     /// <inheritdoc />
     public bool Evaluate(in Context context)
     {
         // cannot use LINQ since 'in' can't be used in lambdas
-        foreach (var c in Conditions)
+        foreach (var c in Terms)
         {
             if (c.Value.Evaluate(in context))
                 return true;
