+++ LethalAchievements/Config/Context.cs
@@ -2,6 +2,10 @@
 
 namespace LethalAchievements.Config;
 
+/// <summary>
+///     The context in which an achievement is being checked.
+///     This is used by <see cref="ICondition"/> and generated by <see cref="ITrigger"/>s.
+/// </summary>
 public struct Context
 {
     /// <summary>
@@ -13,16 +17,6 @@ public struct Context
     ///     The player being checked.
     /// </summary>
     public PlayerControllerB? Player;
-    
-    /// <summary>
-    ///     Creates a copy of this context.
-    /// </summary>
-    public Context Copy()
-    {
-        return new Context {
-            Player = Player
-        };
-    }
  
     /// <summary>
     ///     Creates a context with default values.
